Update sync.md (#24)
Brand new user and I really am not an expert in these matters but I struggled with this, and in the end solved it by removing the final 'espanso'. I'll leave it to you whether you think that is good advice or a quirk of my setup. Great app, thanks.